% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{CLdata}
\alias{CLdata}
\alias{CLdata-package}
\title{Data for the Estimation of Salmonid Abundance in Scottish Rivers}
\description{
Data for the Estimation of Salmonid Abundance in Scottish Rivers
}

